% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_marg.R
\name{cov_marg}
\alias{cov_marg}
\title{Calculates Value of the Marginal-Correlation Function}
\usage{
cov_marg(
  coords = NULL,
  dists.lowertri = NULL,
  nugget,
  cov.model,
  cov.pars.uni,
  log.det.chol = FALSE,
  inv.chol = FALSE
)
}
\description{
Computes a marginal-correlation function,
given the vector of parameters and the distance of their locations.
Different correlation functions are available.
}
