% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovSimpler.R
\name{CovSimpler}
\alias{CovSimpler}
\title{Calculates Value of the Simpler Covariance Function}
\usage{
CovSimpler(
  coords = NULL,
  dists.lowertri = NULL,
  cov.model,
  cov.pars,
  p,
  nugget = 0,
  chol.varcov = FALSE,
  SigmaB = NULL
)
}
\arguments{
\item{varcov}{the covariance matrix.}

\item{log.det.chol.varcov}{the logarithmic determinant of the Cholesky decomposition of the covariance matrix.}

\item{func.decomp}{algorithm used for the decomposition of the covariance matrix in determinant calculation.}
}
\description{
Computes the Simpler covariance function associated to a p-dimensional
multivariate Gaussian random field.
Different correlation functions are available to model marginal-covariance behaviors.
When \eqn{p = 1} the function reduces to the marginal-covariance function (univariate case).
}
