% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSimpler.R
\name{FitSimpler}
\alias{FitSimpler}
\title{Calculates the log-likelihood value for the Simpler Model}
\usage{
FitSimpler(
  data,
  dist.matrix,
  cov.model,
  ini.cov.pars,
  SigmaB = NULL,
  method,
  hessian,
  logpars
)
}
\description{
Computes the value of the log-likelihood for the
Simpler Covariance Model associated to a p-dimensional
multivariate Gaussian random field.
}
