% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grfSimpler.R
\name{grfSimpler}
\alias{grfSimpler}
\title{Simulation of Gaussian random fields with a Simpler Covariance Function}
\usage{
grfSimpler(
  n,
  coords = NULL,
  nx,
  ny,
  xlims = c(0, 1),
  ylims = c(0, 1),
  nsim = 1,
  cov.model = "exp",
  cov.pars,
  nugget = rep(0, length(cov.model)),
  p = 1,
  mean = rep(0, p),
  SigmaB = NULL
)
}
\description{
Simulation of Gaussian random fields with a Simpler Covariance Function
}
