% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikSimpler.R
\name{loglikSimpler}
\alias{loglikSimpler}
\title{Calculates the log-likelihood value for the Simpler Model}
\usage{
loglikSimpler(dist.matrix, cov.model, cov.pars, data, SigmaB = NA)
}
\description{
Computes the value of the log-likelihood for the
Simpler Covariance Model associated to a p-dimensional
multivariate Gaussian random field.
}
