% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikSimpler_optim.R
\name{loglikSimpler_optim}
\alias{loglikSimpler_optim}
\title{Calculates the log-likelihood value for the Simpler Model}
\usage{
loglikSimpler_optim(
  par,
  data,
  dist.matrix,
  cov.model,
  nloc,
  ncov.model,
  nparam,
  logpars = FALSE
)
}
\description{
Computes the value of the log-likelihood for the
Simpler Covariance Model associated to a p-dimensional
multivariate Gaussian random field.
}
