% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikSimpler_optim_uni.R,
%   R/loglikSimpler_optim_uni_externa.R
\name{loglikSimpler_optim_uni}
\alias{loglikSimpler_optim_uni}
\title{Calculates the log-likelihood value for the Simpler Model}
\usage{
loglikSimpler_optim_uni(
  data,
  coords,
  dist.matrix,
  cov.pars,
  nloc,
  cov.model,
  nugget = 0,
  compute.dists = TRUE
)

loglikSimpler_optim_uni(
  data,
  coords,
  dist.matrix,
  cov.pars,
  nloc,
  cov.model,
  nugget = 0,
  compute.dists = TRUE
)
}
\description{
Computes the value of the log-likelihood for the
Simpler Covariance Model associated to a p-dimensional
multivariate Gaussian random field.

Computes the value of the log-likelihood for the
Simpler Covariance Model associated to a p-dimensional
multivariate Gaussian random field.
}
