% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikSimpler_optim_unicov_pmulti.R
\name{loglikSimpler_optim_unicov_pmulti}
\alias{loglikSimpler_optim_unicov_pmulti}
\title{Calculates the log-likelihood value for the Simpler Model}
\usage{
loglikSimpler_optim_unicov_pmulti(
  par,
  data,
  dist.matrix,
  p,
  cov.model,
  nloc,
  nparam,
  logpars = logpars
)
}
\description{
Computes the value of the log-likelihood for the
Simpler Covariance Model associated to a p-dimensional
multivariate Gaussian random field.
}
