% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.FitSimpler.R
\name{lrtest.FitSimpler}
\alias{lrtest.FitSimpler}
\title{Realiza o teste de razão de verossimilhanças para modelos encaixados}
\usage{
lrtest.FitSimpler(fullmodel, reducedmodel)
}
\description{
função que calcula....
}
