% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nparam_covmodel.R
\name{nparam_covmodel}
\alias{nparam_covmodel}
\title{Calculates o número de parâmetros no modelo}
\usage{
nparam_covmodel(cov.model)
}
\description{
Retorna um vetor com o número de parâmetros de cada modelo
da função cov.model e dos parâmetros da função de correlação SigmaB
}
