% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_marg.R
\name{cov_marg}
\alias{cov_marg}
\title{Calculates Value of the Marginal-Correlation Function}
\usage{
cov_marg(dist.matrix, cov.model, cov.pars.uni)
}
\description{
Computes a marginal-correlation function,
given the vector of parameters and the distance of their locations.
Different correlation functions are available.
}
