% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovSimpler.R
\name{CovSimpler}
\alias{CovSimpler}
\title{Calculates Value of the Simpler Covariance Function}
\usage{
CovSimpler(dist.matrix, cov.model, cov.pars, p, SigmaB = NULL)
}
\description{
Computes the Simpler covariance function associated to a p-dimensional
multivariate Gaussian random field.
Different correlation functions are available to model marginal-covariance behaviors.
When \eqn{p = 1} the function reduces to the marginal-covariance function (univariate case).
}
