% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSimpler.R
\name{FitSimpler}
\alias{FitSimpler}
\title{Calculates the log-likelihood value for the Simpler Model}
\usage{
FitSimpler(
  ini.cov.pars,
  data,
  coords,
  cov.model,
  trend = rep("cte", length(cov.model)),
  fix.nugget = FALSE,
  nugget = rep(0, length(cov.model)),
  fix.kappa = TRUE,
  est_mean = F,
  SigmaB = NULL,
  limits = pars.limits(),
  print.pars = FALSE,
  ...
)
}
\description{
Computes the value of the log-likelihood for the
Simpler Covariance Model associated to a p-dimensional
multivariate Gaussian random field.
}
