% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grfSimpler.R
\name{grfSimpler}
\alias{grfSimpler}
\title{Simulation of Gaussian random fields with a Simpler Covariance Function}
\usage{
grfSimpler(n, nsim, cov.model, cov.pars, SigmaB = NULL)
}
\description{
Simulation of Gaussian random fields with a Simpler Covariance Function
}
