% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikSimpler.R
\name{loglikSimpler}
\alias{loglikSimpler}
\title{Calculates the log-likelihood value for the Simpler Model}
\usage{
loglikSimpler(
  data,
  coords = NULL,
  dists.lowertri = NULL,
  cov.pars,
  cov.model,
  nugget = rep(0, length(cov.model)),
  fix.mean = T,
  trend,
  SigmaB = NULL
)
}
\description{
Computes the value of the log-likelihood for the
Simpler Covariance Model associated to a p-dimensional
multivariate Gaussian random field.
}
